Union list

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -2,7 +2,7 @@
 """annotation union list"""
 from typing import List, Union
 
-Vector = List[Union[float, int]]
+Vector = List[Union[int, float]]
 
 
 def sum_mixed_list(mxd_lst: Vector) -> float:
